SPICE update

+++ models/toggle/src/SSA/tau_aux.jl
@@ -1,20 +1,18 @@
 type TauVar
-    neg_nu::Matrix{Int32}
-    #tmp_nu::Matrix{Int64}
+    neg_nu::Matrix{Int}
     g::Vector{Float64}
-    irs::Vector{Int64}
-    dx::Vector{Int32}
+    irs::Vector{Int}
+    dx::Vector{Int}
     l::Vector{Float64}
-    k::Vector{Int64}
+    k::Vector{Int}
 
-    TauVar(neg_nu::Matrix{Int32}, g::Vector{Float64}, irs::Vector{Int64}, dx::Vector{Int32}, l::Vector{Int64}, k::Vector{Int64}) = new(neg_nu, g, irs, dx, l, k)
+    TauVar(neg_nu::Matrix{Int}, g::Vector{Float64}, irs::Vector{Int}, dx::Vector{Int}, l::Vector{Int}, k::Vector{Int}) = new(neg_nu, g, irs, dx, l, k)
 end
 
 function preAllocTauVar(sys::System)
     # useful precomputations
     neg_nu = copy(sys.model.ν)
     neg_nu[sys.model.ν .>= zero(eltype(sys.model.ν))] = zero(eltype(neg_nu))
-    #tmp_nu = sys.model.ν[:, any(sys.model.ν .!= zero(eltype(sys.model.ν)), 1)[:]]
     
     g = zeros(Float64, sys.model.ns)
     for i in eachindex(sys.model.hor)
@@ -24,14 +22,14 @@ function preAllocTauVar(sys::System)
     end
 
     irsBits = vec(any(sys.model.ν .!= zero(eltype(sys.model.ν)), 1))
-    irs = Vector{Int64}(0)
+    irs = Vector{Int}(0)
     for i in eachindex(irsBits)
         irsBits[i] && push!(irs, i)
     end
 
     dx = zeros(eltype(sys.model.x0), sys.model.ns)
-    l = zeros(Int64, sys.model.nr)
-    k = zeros(Int64, sys.model.nr)
+    l = zeros(Int, sys.model.nr)
+    k = zeros(Int, sys.model.nr)
 
     TauVar(neg_nu, g, irs, dx, l, k)
 end
