+++ src/SystemCore/SystemState.jl
@@ -1,13 +1,13 @@
 type SystemState
-    i::Int64
+    i::Int
     θ::Vector{Float64}
     σ2::Matrix{Float64}
     γs::Vector{Float64}
     γmins::Vector{Float64}
-    ns::Vector{Int64}
+    ns::Vector{Int}
     θs::Vector{Float64}
     
-    SystemState(; i::Int64 = 0, θ::Vector{Float64} = Vector{Float64}(), σ2::Matrix{Float64} = Matrix{Float64}(0, 0), γs::Vector{Float64} = Vector{Float64}(), δmin_history::Vector{Float64} = Vector{Float64}(), ns::Vector{Float64} = Vector{Float64}(), θs::Vector{Float64} = Vector{Float64}()) = new(i, θ, σ2, γs, δmin_history, ns, θs)
+    SystemState(; i::Int = 0, θ::Vector{Float64} = Vector{Float64}(), σ2::Matrix{Float64} = Matrix{Float64}(0, 0), γs::Vector{Float64} = Vector{Float64}(), δmin_history::Vector{Float64} = Vector{Float64}(), ns::Vector{Float64} = Vector{Float64}(), θs::Vector{Float64} = Vector{Float64}()) = new(i, θ, σ2, γs, δmin_history, ns, θs)
 end
 
 function Base.show(io::IO, s::SystemState)
